refactor: Create Base inference class to simplify conversion to ONNX, OpenVINO
